+++ src/js/menu/ui/index.js
@@ -4,12 +4,14 @@ import { MenuSection } from "@/components/menuItems";
 import FullscreenVideo from "@/menu/ui/fullscreen-video";
 import SplitChat from "@/menu/ui/split-chat";
 import HideChat from "@/menu/ui/hide-chat";
-import HideVideo from "@/menu/ui/hide-video";
 import HideAvatars from "@/menu/ui/hide-avatars";
 import HideBackground from "@/menu/ui/hide-background";
 import ShowTimestamps from "@/menu/ui/show-timestamps";
 import HideGifSelfie from "@/menu/ui/hide-gif-selfie";
-// import DisableVideo from "@/menu/ui/disable-video";
+
+/*START.NOT_EXT*/
+import HideVideo from "@/menu/ui/hide-video";
+/*END.NOT_EXT*/
 
 const UISection = () => {
   return (
@@ -17,12 +19,13 @@ const UISection = () => {
       <FullscreenVideo />
       <SplitChat />
       <HideChat />
+      {/*START.NOT_EXT*/}
       <HideVideo />
+      {/*END.NOT_EXT*/}
       <HideAvatars />
       <HideBackground />
       <HideGifSelfie />
       <ShowTimestamps />
-      {/* <DisableVideo /> */}
     </MenuSection>
   );
 };